Redo the package move
